fix: Unconnected users displayed as contacts in private accounts searches (#2874)
The way we distinguish between contacts and unconnected users in the search results was broken for private accounts. The teamId used to filter unconnected users belonging to the same team is of the type Option[TeamId], not TeamId, which meant that the comparison searchedUser.teamId == ourTeamId returned true also if both were None.