fix: Generating the conversation name crashes if user data is missing (#2876)
* fix: Generating the conversation name crashes if user data is missing

fixes https://wearezeta.atlassian.net/browse/AN-6923

The code used to generate conversation names assumed that asking for a name with UserId means we have the name in the local storage. It doesn't have to be so if the team is not synced, and then we will have a crash on trying to get the username from Map[UserId, Name] without providing the failsafe.
The PR changes that so now if the map does not have the name for the given userId, we will ignore it.