fix: Race condition on conversation list click while reloading the list
fixes a bug reported in Google Play Console

We use a mutable list for items and in rare cases it may lead to a situation when a click is executed while the list is refreshing.
Rare, but it results in a crash. The ideal solution to this would be to change the items into an immutable list (held in a `var`
instead of `val` instead) but that messes up with how we put items in folders and I don't want to spend time on this right now.
Instead, I wrapped the `onClick` methods in `Try`. This way the worst what will happen is that the user will have to click again.