fix: Bottom dialog in Group Participant View shows only one option (#2879)
fixes https://wearezeta.atlassian.net/browse/AN-6922

ConversationOptionsMenuController implements complex logic to decide what options should be displayed in the bottom dialog. Some of the checks are not straightforward - for example, to check if the bottom dialog is opened on the group or single participant view, it checks if `selectedParticipant` is set, and that's a field which usually is used for something very different. So, if we open Group Participant view, `selectedParticipant` should be `None`, and if open Single Participant view then it should be `Some(<user id of the participant>)`. In one of previous PRs I started to set this field to `Some(...)` even when we opened the Group Participant view, because I thought it was harmless and it made the code a bit simpler.

It would be great to rethink the logic (I'd suggest to split it into a few branches depending directly on what screen is displayed) but for now I just reverted my previous change.