fix: The "Done" button on Create Group Conversation screen unresponsive for large teams (#2880)
fixes https://wearezeta.atlassian.net/browse/AN-6912 (?)

I wasn't able to reproduce this behaviour but from reading the code I guess the performance problem might lie in that we try to "track" how many of added users are guests. This information is then ignored because we don't use a tracking system anymore. I removed it from the code.

That being said, creating a new conversation is a complex operation and if the device is already under heavy load because of processing a big number of team members, it can take some considerable time.