refactor: Remove the Generators class
A long time ago we thought about using `org.scalacheck.Arbitrary` everywhere in our tests. We wrote a huge utility class,
`Generators`, and then it turned out that it's much easier to write unit tests without it. A few changes to existing unit
tests are needed, but other than that we can safely remove `Generators`, making the codebase smaller, and test compilation
faster.