fix: Unable to open Group Participants View after removing a user
fixes https://wearezeta.atlassian.net/browse/AN-6934

I made the `selectedParticipant` signal update to `None` on removing the seleced user from the conversation and also on
explicitely opening the Group Conversation View (as sometimes we can open the Group view while having the participant
selected - this is not a bug, but may lead to some misunderstandings).
Also, I changed the `otherParticipantId` signal to such that it will never point to a user removed from the conversation.
Just to be sure.