fix: Update team data (#2883)
fixes https://wearezeta.atlassian.net/browse/AN-6857

The team data is now updated in real time on receiving an update event - until now the data in storage was updated, but UI displayed it with some delay. Also, because in case of large teams the event may not come, now every time we go to the profile, the app refreshes
team data. It had to be done with a separate call than syncTeam, because syncTeam also retrieved members and roles.