feat: File restrictions (#2894)
In a custom build we can now have an enabled whitelist of file extensions. If we have it, only files with extensions on the list are allowed to be send and received. The list is disabled on the main build.

https://wearezeta.atlassian.net/browse/PROD-55

The PR contains also a partial fix for bug where the user is not able to share files from another app if the file is a raw file from a non-standard directory. The fix makes it possible to go through the "share file" screen but the bug still prevents us from actually sending the file.