fix: Remove file restrictions on file pickers (#2897)
When we open a file picker from inside Wire, and try to force it to hide some of the files basing on their extensions,
some file pickers end up restricking access to whole folders (e.g. Huawei phones, Android 8.0). Until we find out
how to handle it better, we decided not to restrict the file picker. The user will be allowed to choose any file,
and only then Wire will prevent the file from being sent. The code that adds the list of allowed MIME types to the intent
will stay commented out in case we change this decision.

I also moved the dialog with the error message a bit to make sure it is displayed (there was a report that it doesn't
appear on one device). And I fixed a bug when the error system message was displayed with empty text when the extension is unknown (rare, but possible).