Wire Signals library (#2901)
This PR replaces the signals functionality from Wire Android codebase with signals from the separate library: https://github.com/wireapp/wire-signals

Signals, event streams, CancellableFuture, and classes aiding them, were moved to the library.
    
The Threading class was split in two. The basic functionality is in the library, the parts specific to Wire Android (e.g. the Ui thread) is still in Wire Android. One of the next steps in developing wire-signals aims to make the library part easier to use, so that the Wire Android part can be simplified and then use as an example how wire-signals can be used in other projects.
Some of the flaky unit tests were fixed.