fix: Avoid displaying 'Default' as the username in system messages (#2906)
fix: Avoid displaying 'Default' as the username in system messages

fixes https://wearezeta.atlassian.net/browse/AN-6873
fixes https://wearezeta.atlassian.net/browse/AN-6859

In some cases, when the user is removed from the team or the conversation, we may not know their name and "Default" is displayed in the system message. To avoid that, I addeed "sync if needed" to the MemberLeave event handling, just as it is done for MemberJoin. I also added new strings which are used if we still don't know the name of the user even after syncing (might happen for example when the user is removed from the team).