fix: Team data gets updated but new team data is not stored (#2907)
fixes https://wearezeta.atlassian.net/browse/AN-6857

I introduced this bug a few weeks ago when I was working on team data updates - I removed the possibility
to receive a completely new team data. This PR fixes that. The side-effect, which led to discovering the bug,
was that the team name didn't appear on the profile view for new team members.