New backups: Encryption
EncryptionHandler restored as a new PR.

I rewrote this code as closesy as possible from Scala. Please take a look at com.waz.service.backup.EncryptionBackupHeader
and com.waz.utils.crypto.LibSodiumUtils. We know that code works, but it's complex and hard to read. I'd suggest not to
refactor it until we are sure our changes won't break anything.