New Backups (17+3e): Encryption and Decryption (#2994)
* New backups: Put together backup creation in the use case

* New backups: Metadata handler

* New Backups: Create a backup from UI

This PR contains code connecting Scala and Kotlin (1) and of a fix to BackUpDataSource<T, E> (2)

1. In BackupExportView in Scala I switched from calling the old backup functionality in BackupManager to calling the new functionality through KotlinServices and BackUpViewModel. This is by necessity a bit hacky: KotlinServices provides a way to call Kotlin from Scala, but it's not a fragment or an activity, so injecting BackUpViewModel is done through a small additional Koin library: koin-java. Its official purpose is to enable calling Koin from Java, but it turns out it also greatly simplifies calling Koin from Kotlin where standard methods don't work.

BackUpViewModel handles CreateBackUpUseCase and in the future it should also handle RestoreBackUpUseCase. It hides from KotlinServices (and Scala) the fact that we use coroutines. Instead, Scala provides two callbacks: onSuccess and onFailure. One (and only one) of them should be called when the Kotlin side finishes working.

2.
The bug was that the abstract type BackUpRepository<List<File>> was referred to and treated as BackUpRepository<File> in CreateBackUpUseCase. The compiler didn't catch it, and on top of that Koin was happy to inject the repositories even though their types didn't fit. Only when I started to test manually, the app crashed with a ClassCastException. Apart from fixing the type, I had to add names to arguments in Koin. It was necessary also for the point #1 to let Koin know how to create BackUpViewModel.

I made also some changes to the how we write json files and then zip them, but this part will be changed once again. Here I just wanted to make sure it works.

New backups: Encryption
Rewriting encryption/decryption back to Sodium v1