New backups: Remove the popup and dead code from the Scala side
The order of steps in initialization changed a bit. Previousy, the backup was being restored as the first step
in `accountsService.createAccountManager`. We were initializing the user's account already having all data from
the backup, and afterwards we were performing a "cleanup" - resetting certain preferences which triggered
initial sync. Now, the user's account is created empty and only then we restore the backup. The cleanup had to
be moved to afterwards the backup restoration.