Bump Android SDK 29
Work in progress.

Two main functionalities that were broken by this bump are sharing assets and the camera. The camera can probably be rewritten using the new API and we just need some time.
The assets are a bit more tricky. Until now we relied on URIs to direct files. The proper way in SDK 29 would be to use `FileDescriptor`s instead, but that would mean
a cascade of changes in many other places in our code. And I think we would still need to use the old way for older Androids. I'm tryin to find a middle solution.