Allow for custom HTTP Proxy settings in AndroidManifest.xml
fixes https://wearezeta.atlassian.net/browse/AN-7096

Our customer wants to be able to change the proxy settings in an already compiled APK.
Currently the proxy settings are hardcoded (taken from default.json or custom.json and put as constants in the BuildConfig class).
However, the customer is able to open the APK, edit AndroidManifest.xml, and put it back. Then, when launching the application,
we need to check if there are valid proxy details in AndroidManifest.xml, and if yes, use them instead of the ones from BuildConfig.

I did my best to make the code look good, but this is all pretty hacky from the very start. On top of the hackiness, we have two
implementations of the proxy details: one in Scala, one i Kotlin. I decided not to connect them, but rather write the code twice.
The proxy details are set early in the initialization and I feared that trying to make Scala use Kotlin at this point, or vice versa,
may lead to unreadable code and even introduce some bugs.
On the other hand, I tried to make both implementations as similar as possible. This should help with readability.