fix: TrackingService methods crash if tracking is disabled
Methods of the new TrackingService implementation used `isTrackingEnabled.head.collect { case true => ... }` to check
if the tracking was enabled, and if yes, then to perform some actions. What should supposed to happen if tracking was
disabled was not specified. In such cases, Future.collect failed with `NoSuchElementException` even though in all places
where those methods were used, we expected the successful result (of the type Unit).

This mistake is easy to make because in collections, e.g. `Seq.collect`, if the case for `collect` is not specified,
we will simply get an empty collection. Also, if we use `Future.collect` but expect `Unit` in return, it's ok to allow
`collect` to fail. But if we expect `Future[Unit]` we need to use `Future.map` or `Future.flatMap` and specify what
should happen in all possible cases.

I introduced a special private method `withTracking` which hides these details and requires only the implementation of
what should happen if tracking is enabled. It also prevents the future from failing from other reasons (since this is
TrackingService I assumed that even if something fails in it, it shouldn't crash the app).

I looked through the rest of the code and changed from `Future.collect` to `Future.foreach` and `Future.map` in a few
more cases.