Allow for custom HTTP Proxy settings in AndroidManifest.xml (#3037)
* Allow for custom HTTP Proxy settings in AndroidManifest.xml

fixes https://wearezeta.atlassian.net/browse/AN-7096

Our customer wants to be able to change the proxy settings in an already compiled APK.
Currently the proxy settings are hardcoded (taken from default.json or custom.json and put as constants in the BuildConfig class). However, the customer is able to open the APK, edit AndroidManifest.xml, and put it back. Then, when launching the application, we need to check if there are valid proxy details in AndroidManifest.xml, and if yes, use them instead of the ones from BuildConfig.