Bump Android SDK 29 (#3023)
* Bump Android SDK 29

Two main functionalities that were broken by this bump are sharing assets and the camera. The camera can probably be rewritten using the new API and we just need some time. The assets are a bit more tricky. Until now we relied on URIs to direct files. The proper way in SDK 29 would be to use `FileDescriptor`s instead, but that would mean a cascade of changes in many other places in our code. And I think we would still need to use the old way for older Androids.

* move strings to constants in ShareActivity.getPath
* Camera to Camera2 API Migration  (#3034)

Co-authored-by: Maciej Gorywoda <makingthematrix@gmail.com>
Co-authored-by: BradleyWire <57259490+BradleyWire@users.noreply.github.com>