fix: Get random profile pictures from Unsplash (#3081)
* fix: Get random profile pictures from Unsplash

fixes https://wearezeta.atlassian.net/browse/AN-7047

I'm not exactly sure how long this functionality was broken but it looks like it was for a long time. The previous code
tried to treat an Unsplash URL as if it was a regular asset that was supposed to be downloaded through our backend.
The backend failed because it didn't understand the request.I rewrote `AccountManager.addUnsplashPicture()` and added methods to `AssetService` and `AssetClient` which download the picture from Unsplash. Then the picture is set as the profile one and sent to the backend exactly as if it was a regular profile picture.

I also added a button in Dev Settings to test it. From now on it can be used if we want to test any change in code connected to profile pictures.

* changes requested in the review