fix: The password dialog disappears when another person calls
fixes https://wearezeta.atlassian.net/browse/SQCORE-46

1. You have the password dialog enabled, either because you're on a custom build, or you enabled it in the options.
   It appears when you open the app or if it was in the background for more than 10s (or 30s dependent on the build).
2. On new phones there's also a biometric check but that's up to Android.
3. If someone calls, the call screen appears on top of the password dialog. The user can answer the call.
   That shouldn't be allowed. The user should have to enter the password first.
4. And also, when the call ends, or is dropped by the caller, the user goes back to the conversation list and doesn't
   see the password dialog anymore.

I fixed that. Now we have another password dialog on the call screen (so the user has to type in the password to answer
the call - that also didn't work) and the password dialog on the main screen doesn't go away.

That however created a glitch. The two password dialogs don't communicate with one another. So if the user types in
the password on the call screen, after the call ends, they need to type in the password again on the main screen.
Some decisions were taken when we implemented the calling long time ago and connecting them now proves quite difficult.
I keep running into situations when again the password dialog is gone in some corner cases.
In the end, I decided to leave the glitch. There is a plan to rewrite the calling activity code. Maybe then it will be
possible to fix this problem in an easy way.