fix: Using one of three audio filters crashes the app
fixes https://wearezeta.atlassian.net/browse/SQCORE-74

This is work in progress. So far investigations shows that the crash happens inside the AVS native method
`AVS.applyEffectPCM(String, String, int, int, boolean)` and it's not an exception being thrown but an immediate crash:
```
libc: Fatal signal 11 (SIGSEGV), code 2 (SEGV_ACCERR), fault addr 0x77cfcb2e60 in tid 28744 (waz.zclient.dev), pid 28744 (waz.zclient.dev)
```

It happens only for some values of the fourth argument - the id of the audio effect - but not for others.
Now I'm waiting for the opinion of the AVS team on this.

I also used this opportunity to delete a lot of old unused code left after we rewrote the audio recording in Kotlin.
There's so much of it that I decided to already make a PR.