fix: Don't ask for the password when initiating the call
A fix to a problem which came up after the fix to the password dialog security issue. Generally this is a lesson to me
to write myself a small list of possible bugs I can introduce when changing a piece of code so I can make some manual
tests (or write unit tests when possible) before I make a PR.

The problem was that now the password dialog appears also when the user starts a call and when they receive a call but
after they typed in the password already. This should not be the case. This PR fixes that.

Note that it does not fix some cases when a new password dialog is displayed on top of an older one and the user has to
type in the password twice. This will be addressed in a later PR.