feat: Shared tracking id
implements https://wearezeta.atlassian.net/browse/SQCORE-121

The tracking ids are already present in the client code, but now every device has its own tracking id and they are not connected.
This PR introduces sharing tracking ids among devices belonging to the same user. The general idea is that every time a new device
is added to the user, that device generates a new tracking id and sends it to all other devices. This happens also when the device
updates for the first time to the version containing this feature. We are aware there is a risk of race condition when two devices
send their tracking ids in the same time, but we decided it's small enough and we can ignore it.

The change requires an update in generic-message-proto. The new version is 1.27.1. It contains a new field called DataTransfer and
inside it there's a field trackingIdentifier which is the trackingId that should replace the local tracking id. In the future,
DataTransfer might contain more fields, but for now it's only this one, so in the Scala code I made a shortcut that let us easier
unpack the tracking id from the incoming message.

I also added a new button to the Dev Settings that can be used to generate a new tracking id without the need to remove and re-add
a device. This can help in testing.