fix: A "unit" test in AssetServiceSpec was trying to connect to the backend
The problem was discovered because of a bug introduced on the backend in the code generating SHA sums.
A test in AssetServiceSpec which tests SHA sums started to fail, even though it was supposed to be an unit test
and it shouldn't have nothing to do with the backend.

The test is quite complex. It's a component test, not unit test, but I think it has its value - it tests a logic path
through an important part of AssetService functionality. I rewrote parts of it which used real encryption and AssetClient.
Now these are mocked.