SSO password lock
fixes https://wearezeta.atlassian.net/browse/SQCORE-170

An SSO user does not have a login password, and so they can't use the app lock functionality. This PR solves it by introducing a special password,
only for SSO users, which is stored - hashed - in user preferences. There is a new dialog which opens after the login if the app lock is enabled
and the password is not set. Later on, that password is used in all places where for a standard user the login password would be used.
The password can be cleared only by deleting local data. The "passcode" switch in the Options was changed to a new view which now contains the switch
and a new button that the SSO user can use to change their password. For a standard user this button is invisible.

I use the Sodium library for hashing. I decided to move around the code for this so I will be able to write unit tests with "dummy" hashing implementation.

I still work on the UI of the new password dialog as well. The logic should stay the same and you can already review it.Fix for redundant password dialogs