wire-signals bump to 0.3.1 (#3098)
This PR consists of the bumping the version and two very minor changes to the code required by changes in wire-signals. The wire-signals update is mostly about making it work with Scala 2.12 and 2.13 - some code was specific to Scala 2.11 we use in wire-android and had to be changed. In 0.3.1 wire-signals compiles under all three versions.

There is also one new method, onUpdated, in the Signal class. It works like onChanged but returns not only the new changed value, but also the old one, so they can be compared to see what exactly changed from one to another. This is a pattern that comes up regularly in the wire-android code - we not only want to know there was a change but also what was changed - and until now it had to be solved in wire-android. This method should simplify our code in a few places.
(But it's not tested yet - I want to first rewrite unit tests in mUnit).