fix: Video recording on Android 11 doesn't work
fixes https://wearezeta.atlassian.net/browse/SQCORE-219

Starting Android 11, Android does not allow a third-party app to use `intent.resolveActivity` to easily learn
about other installed apps. One way to solve it is to simply stop using `intent.resolveActivity` and just attempt
to start a new activity and deal with the `ActivityNotFoundException` if it so happens that the app is unable
to find an app for the given intent. In this case, we want to open a video recording app and we can safely
assume that each device we support has it. In a few other places in the coded I decided to use the new code
to guard us from the `ActivityNotFoundException` as well.

I also fixed one trivial bug in `CursorController.toggleEphemeralMode` where two objects of different types
were compared. The fix shouldn't have any practical effect.