fix: The visibility of the "eye" icon on the "Change Passcode" dialog (#3112)
The "Change Passcode" dialog is always opened from the Settings. Settings ignore the theme from ThemeController -
it's always dark theme in the settings. It means that if we open the SSO password dialog from the Settings
(i.e. the "Change Passcode" case) we need to always use the dark mode as well or otherwise the "eye" icon will
be displayed black instead of white and it will be hardly visible.