fix: NetworkModeService crashes on Android 11
This fix uses a network callback to listen to changes in the network type. It avoids using NetworkInfor and ConnectivityManager.getActiveNetwork.
The only permission we need is ACCESS_NETWORK_STATE. We ask for it when the user opens the app for the first time after the update or installation,
just before we register the network callback.

Instead of many types of cellular network we now have only one mode: CELLULAR. Because of this I removed code which warned the user that the network
is slow. It checked if we are on 2G and EDGE networks. Anyway, it was only a partial solution - it was still possible for the user to be on 3G or 4G
and have poor bandwidth.

There is an old method in NetworkModeService: `override def isDeviceIdleMode: Boolean = false`. I removed comments around it, but left it as it is
for this fix. I want to change it to something more meaningful in another PR.