fix: File sharing
I rewrote ShareActivity and made some changes to AssetService. The rest is some slight refactoring.
I decided that the best way to simplify and make sure that the widest range of devices will work with this code is to copy
files temporarily to our cache directory and then sending that copy. This makes it easier for us to navigate Android
paths and permissions. Unfortunately, doing it for videos would be an overkill. Instead, in accordance to how we handle videos
in other places in the assets functionality, we should access them directly. The side-effect of this is that on the sender's side
the app may sometimes be forbidden from accessing the file - either at the moment of sharing the file or when the sender wants
to play the file which is already displayed in the conversation. I tried to fix that but it was a bit of a trial and error process.
I can't be sure that it works on all devices.

If you have time, please download this PR, install it on your phones, and experiment a bit.

There's a place in AssetService where we validate the file by computing the sha and comparing it with the one we already have.
I decided to drop this for videos - it takes too long and I suspect the app may reach a timeout at one of the internal steps.
I'm thinking about dropping validation also for other file types if they are too big.