feat: Feature Flag App Lock
https://wearezeta.atlassian.net/browse/SQSERVICES-58

Show App-lock feature setting in Wire account setting based on App-lock feature status.
The feature requires communication with a new endpoint for checking the feature flag status  every time when the app goes to the foreground.
If the flag is enabled, we use the returned config to check if the app lock is enforced and what is the timeout.
Contrary to the "forced" setting in the build config, if the app lock is forced by the feature flag, we don't wipe the data clean if the user fails to validate.
If the flag is disabled, local settings apply, just as until now.
The feature flags work only for team accounts. If the account is personal, we act as if the flag was disabled.
The PR also contains some refactoring - I moved the AppLockEnabled preference from GlobalPreferences to UserPreferences, I created user preferences
for the "forced" setting and for the timeout, and I created a temporary flag in GlobalPreferences to migrate AppLockEnabled: If the user had AppLockEnabled set to true
until now, that setting should be copied to UserPreferences.