Changes required to make wire-android work with wire-signals 0.3.2 (#3142)
To make Wire Signals code compatible across Scala versions (2.11, 2.12, 2.13) I had to change one thing in the API:
RefreshingSignal and AggregatingSignal take as an argument a reference to a "loader" - a future which will initially
compute the value of the signal. Until now, that reference was given by name, which meant that it could be use many
times, lazily. It wasn't called in the moment when it was given to the constructor, but it was saved in the signal
and called later, as many times as needed. This is no longer the case. Now, the same functionality is achieved by
giving to the constructor, as an argument, a function which will produce the future each time it is called.

It's also more intuitive this way. The old way gave an impression that the future is called once, immediately.

The changes in Wire Signals 0.3.2 include also fixing a bug in EventStream.mapAsync, and some compilation-time optimizations.