feat: Allow session reset from the decryption error message (#3118)
* feat: Allow session reset from the decryption error message

see https://wearezeta.atlassian.net/browse/SQCORE-30

When the user sees an error message that the decryption failed, they should be able to fix that simply by clicking on it, instead of having to go the other user's profile and reset the session manually from there to every device of that other user.

The feature consists of:
 - a changed error message,
 - a new message for a fixed error,
 - the functionality that triggers session reset for all error messages from the given device (no matter in what conversation),
 - a spinner appearing in the centre of the screen while the reset is in progress,
 - a change to MessageData and MessageDataDao so that from now on error messages will carry the error code and the sender device id,
 - a migration,

I also used this opportunity to refactor the code a bit. Message view parts are old code - I wanted to improve its readibility. And I found that there is some unused code left after the recent changes in audio/video functionality.

* changes requested in the new Design Doc + refactoring
* the session reset system message
* Removing old migration tests
* removing unused import
* A toggle for replacing messages with fake errors
* fixing unit tests