App Lock Feature Flag #2 (#3151)
* Partial migration from the account to the custom password for App Lock

This is a follow-up of introducing the App Lock Feature Flag. Until now we used the account password for the app lock, and only if the user was SSO (and so, no password) we used the custom one. Now we will migrate to using the custom password in the app lock for all users. For some time now, after the user updates to the new Wire Android version, they will still use the account password for the app lock (nothing changes from the user's perspective) but if they switch the app lock off and on again, or if it is forced by the feature flag, we will display a dialog asking the user to create a custom password.

This feature also affected others a bit - removing old devices, and logging in. We need to provide the account password to log in, as well as to remove a device, but until now the code relied a bit on PasswordController which now is controlling only the custom password. I had to change it, and also I removed a bit of code which I think shouldn't be there - onPasswordReset was called for apparently no good reason. I want to take a look at it once again, before merging.

I call the custom password "custom", instead of "app lock password", because I suspect in the future it can be used
for other purposes as well. But I left as they were the names of the user preferences used to store the encrypted
custom password: sso_password and sso_password_iv. Renaming them would mean yet one more migration.