fix: Un-throttling the networkMode signal in WebSocketService. (#3159)
fixes https://wearezeta.atlassian.net/browse/SQCORE-420 (ANR because of WebSocketService)
fixes https://wearezeta.atlassian.net/browse/SQCORE-359 ("No internet connection" bar constantly displayed)
connected to https://github.com/wireapp/wire-signals/issues/47

SQRCORE-359 is there because of the bug in wire-signals (linked above), but SQCORE-420 would still be there,
even if that bug was already fixed. SQCORE-420 happens because with the throttled networkMode there is too much
delay between Context.startForegroundService() and the consecutive Service.startForeground(), and Android kills
the service after a timeout. To fix this, I removed the throttling from networkMode and made some changes in
WebSocketService to ensure that startForeground is called a bit quicker.

Previously, I wanted to use throttling to prevent situations, where the network changes rapidly and that causes
innecessary reactions in the app (usually when the connection quality is low). I may think of reintroducing it
when I fix the underlying bug in wire-signals (but then I will not use it for WebSocketService).