fix: Perform full sync on backup restoration and delete conversations missing on the backend (#3162)
fixes https://wearezeta.atlassian.net/browse/SQCORE-127

When the user deletes a conversation between the backup and its restoration, the restoration brings back the deleted conversation. The fix performs a full sync after the restoration and during syncing conversations it compares the results from the backend with what's in the storage (after the restoration). Any conversation that is not present on the backend will be deleted, together with its members, messages, etc.
On top of it, the full sync applies updates on top of the restored backup - that was also missing from the implementation.