fix: An unrecognized event crashes the decryption process (#3163)
This is a fix to a problem reported on the Internal/staging setup, but the problem might occassionally arise on production as well. If an event can't be deserialized in ConversationEventDecoder, it is wrapped in UnknownConvEvent. But then, in PushService.processEncryptedRows, we assume each event is properly decrypted and we use an unsafe operation to unwrap it. For an unrecognized event this results in a crash of the decryption process, nad since the event is not removed from the queue, when the process restarts, the event is still first in the queue, and crashes the process again. The user experiences it as if no new messages were coming at all.

Usually it shouldn't happen - we should be able to recognize all events. But there is no point in using the unsafe operation if it can be easily avoided. I changed that part of the code to the one that doesn't crash on an unrecognized event, and I added log statements, so we can learn what was it that couldn't be deserialized.