fix: Backup restoration freezes the app
fixes https://wearezeta.atlassian.net/browse/SQCORE-450

A line of code introduced in September deletes user preferences which are used for user verification on init. While the app is on,
everything works fine - that's why it was difficult to track the bug.
I think I copied that line from the previous implementation of backups but in the new one it does not serve any purpose.