fix: NullPointerException in markdown
Google Play Console reports crashes because of an NPE in rendering markdown, but I can't reproduce it
and the stacktrace is not very helpful. I decided to use the same solution as with an earlier occurences
of ArrayIndexOutOfBoundsException - that is, let's catch the exception, log it, and if we're on production,
simply allow the markdown to fail, but don't crash the app.

Hopefully, this way we will eventually get more information from an Internal build which will reproduce the bug.