fix: Optimize push notifications
In Google Play Console we get constant ANRs about that we didn't manage to process an event in time in WebSocketService
("Context.startForegroundService() did not then call Service.startForeground()"). Also, we get reports a lot of reports
about that new messages come only after the user opens the app - which means that FCM messages are also not received or
not processed in time.

I have read through the code for both web socket and FCM events/notifications and made some optimizations. (And also
it's possible I fixed something with the default notification channel for Android 8.0+ that wasn't working before).
I'm not sure about the results, but it shouldn't be worse than before.