Add the possibility to export conversations as ZIP (XML+mediadata+viewer)
Add the possibility to export conversations as ZIP-file containing the
conversation as an XML-file.
It is possible to define the desired period of time and whether
media-files, profile pictures and the HTML-Viewer should be included in
the export.
The HTML-Viewer enables the user to view the exported data directly.

This function was developed in connection with Patrick Nolte's
bachelor thesis at the Institute for Networks and Security at
the Johannes Kepler University Linz.