refactor: Use new protobuf 1.28.2 (#3193)
* refactor: Use new protobuf 1.28.1

I know this is huge.

The old protobuf implementation used the now obsolete "nano" mode which created very small Java classes which had to be accessed and modified directly, i.e. every field was public and it was up to us to ensure we use it in the right way. "nano" is no longer used in protobuf, so when I updated protobuf, I had to change to the more standard "lite" implementation. "lite" uses getters and for modification it has builder classes. That changed a lot how the protobuf classes need to be handled in our `com.waz.model._` classes, esp. `GenericMessage` and `GenericContent`. On top of that, we recently merged two protobuf libraries: generic-message-proto and backend-api-proto. I had to make so many changes that instead of trying to contain them and change as little as possible (not possible) I decided to refactor also the code which touched those classes, so that it all works better together.

Before, we used a lot of implicit objects with `apply` and `unapply` methods for convertions. In many places it meant that we converted the proto data into logic case classes, used them for checking something, then dismissed them, and then we made convertions again when we needed them again. I removed those implicit objects and introduced "newtypes" - wrapper case classes around proto data. Each of them has a `lazy val unpack`, equivalent to the previous `unapply`, but this time after the first convertion, the converted ("unpacked") version is remembered so we don't get the performance penalty when we need to mae another convertion. Instead, I guess the app will now need a bit more memory. We need to test it.

Also, I made some performance optimizations in `GenericMessageService`. Until now with every new event we created and then left for garbage collection a lot of intermediate collections. I decided it happens so often that it might make sense to use one mutable cache instead. This too needs to be tested. Fortunately, both `unpack` and this can be easily converted to versions which will be slower but maybe safer, so if we find a bug in this implementation,
we may try it.

* Changes to WebLinkPartView
* trying to resolve dependencies conflicts around protobuf
* Removing dependency conflicts in the protobuf library
* changes requested in the code review