A revert of a part of a code changed in "Optimize push notifications (#3180)" (#3216)
* A revert of a part of a code changed in "Optimize push notifications (#3180)"

Since 3.65 which introduced #3180 we get reports from Google Play Console about a crash connected to notifications:
https://play.google.com/console/u/3/developers/7098984309886892484/app/4973241010395499500/vitals/crashes/a5a7d5b2/details?installedFrom=PLAY_STORE&days=30

Unfortunately it doesn't say much. From when the crash started happening and from what I found in Google, it seems to be a problem with notification channels and/or with that too much time passes between when the notification comes to the device and when the app is displayed in the foreground. So, that would mean, the crash may happen when the app is in background, when it is killed, or when the device is sleeping.

I reverted two parts of the code that may have caused it. One, in WebSocketService, creates a new notification channel (before I thought we already have all channels created and this is not needed, but maybe it is) and the other, in FCMHandlerService, schedules a fetch from the backend instead of doing it immediately (maybe doing it immediately is too heavy and the app is killed).

Notice both "maybe". This is a guesswork. I simply don't see any other change that could cause this problem.

* reverting a revert in FCMHandlerService