fix: Another try at fixing the notification channel crash from Google Play Console
I looked through the code of the commit which most probably causes the crashes and through Stack Overflow entries on similar issues.
I found this: https://stackoverflow.com/a/51378281/2975925 which says that we shouldn't use localized strings as the notification
channel ids, and I followed this advise in WebSocketService, and also I decided to bring back a line that ensures the notification
channel is created when the app goes to foreground (just as it was before).

We are still not able to reproduce the crash 100% but I think this PR should help.